+++ examples/streamlisten.rs
@@ -3,16 +3,15 @@
 // and multiple threads.
 //
 use futures::StreamExt;
-use pcap::stream::{PacketCodec, PacketStream};
-use pcap::{Active, Capture, Device, Error, Packet};
+use pcap::{Active, Capture, Device, Error, Packet, PacketCodec, PacketStream};
 
 pub struct SimpleDumpCodec;
 
 impl PacketCodec for SimpleDumpCodec {
-    type Type = String;
+    type Item = String;
 
-    fn decode(&mut self, packet: Packet) -> Result<Self::Type, Error> {
-        Ok(format!("{:?}", packet))
+    fn decode(&mut self, packet: Packet) -> Self::Item {
+        format!("{:?}", packet)
     }
 }