+++ src/stream.rs
@@ -4,8 +4,8 @@
 use super::Activated;
 use super::Capture;
 use super::Error;
-use super::Packet;
 use super::SelectableCapture;
+use crate::PacketCodec;
 use futures::ready;
 use std::io;
 use std::marker::Unpin;
@@ -13,11 +13,7 @@ use std::pin::Pin;
 use std::task::{self, Poll};
 use tokio::io::unix::AsyncFd;
 
-pub trait PacketCodec {
-    type Type;
-    fn decode(&mut self, packet: Packet) -> Result<Self::Type, Error>;
-}
-
+/// Implement Stream for async use of pcap
 pub struct PacketStream<T: Activated + ?Sized, C> {
     inner: AsyncFd<SelectableCapture<T>>,
     codec: C,
@@ -43,10 +39,12 @@ impl<T: Activated + ?Sized, C> PacketStream<T, C> {
 impl<T: Activated + ?Sized, C> Unpin for PacketStream<T, C> {}
 
 impl<T: Activated + ?Sized, C: PacketCodec> futures::Stream for PacketStream<T, C> {
-    type Item = Result<C::Type, Error>;
+    type Item = Result<C::Item, Error>;
+
     fn poll_next(self: Pin<&mut Self>, cx: &mut task::Context) -> Poll<Option<Self::Item>> {
         let stream = Pin::into_inner(self);
         let codec = &mut stream.codec;
+
         loop {
             let mut guard = ready!(stream.inner.poll_read_ready_mut(cx))?;
             match guard.try_io(|inner| match inner.get_mut().inner.next() {
@@ -55,8 +53,7 @@ impl<T: Activated + ?Sized, C: PacketCodec> futures::Stream for PacketStream<T, 
                 Err(e) => Ok(Err(e)),
             }) {
                 Ok(result) => {
-                    let frame_result = result.unwrap()?;
-                    return Poll::Ready(Some(frame_result));
+                    return Poll::Ready(Some(result?));
                 }
                 Err(_would_block) => continue,
             }